+++ shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./roles";
+export * from "./extraction";
 export * from "./league";
 export * from "./valuation";
 export * from "./player";